KANBAN-1322 flatten ontologyBrowser/browse/ into ontologyBrowser/
Sethrcamp's review noted the browse/ subfolder isn't doing anything
since ontologyBrowser/ has no other contents. Moves the 10 files up
one level and rewires relative parent paths and the routes.jsx
import accordingly.
