+++ src/containers/ontologyBrowser/style.module.scss
@@ -1,4 +1,4 @@
-@import '../../../theme.module';
+@import '../../theme.module';
 @import '~bootstrap/scss/functions';
 @import '~bootstrap/scss/variables';
 @import '~bootstrap/scss/mixins';
